コンソールクライアントからのチームとグループ登録への対応 (#174)
* api/admin/create_group を api/admin/submit_group に変更．関連するメソッド類もリネーム．

* admin_upload_team.py を api_team_uploader.py へリネーム

* group_uploader/submit.* のコマンドラインオプションを修正

* group_uploaderのconfigに.envを読み取る処理を追加

* api/admin/create_group エンドポイントを改めて追加

* チームアーカイブのパスが存在しない場合は削除処理をスキップする条件を追加

* group_uploaderに.env.exampleを追加

* adminのAPIチームアップロードでもバージョン文字列が空の場合は日時を自動設定するように修正

* チーム名にアンダーバーを許容するように変更

+++ server/team/views.py
@@ -81,11 +81,12 @@ def delete_team(team_id):
             return redirect(url_for("team.index"))
 
         # Delete the archive file
-        abs_path = os.path.join(current_app.static_folder, os.path.dirname(team.archive_path))
-        if os.path.exists(abs_path):
-            # print(f"Delete {abs_path}")
-            current_app.logger.info(f"detele_team: Delete {abs_path}")
-            shutil.rmtree(abs_path)
+        if team.archive_path is not None and team.archive_path != "":
+            abs_path = os.path.join(current_app.static_folder, os.path.dirname(team.archive_path))
+            if os.path.exists(abs_path):
+                # print(f"Delete {abs_path}")
+                current_app.logger.info(f"detele_team: Delete {abs_path}")
+                shutil.rmtree(abs_path)
 
     db.session.delete(team)
     db.session.commit()
